fix: hash the password near the query
Couldn't mock the user creation function when sending the hashed
password, to fix this, it was decided to send the plain password to the
database function and hash it just before saving it to the database.

OBS. I like this solution better just because the database function
ensures that the password is hashed before saving it. Locallity of
behavior